backend/requirements.txt
asgiref==3.5.2
Django==4.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djoser==2.1.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
PyJWT==2.4.0
pytz==2022.2.1
sqlparse==0.4.2
